[ONNX] Passes to reuse existing 'fake_mode' if possible
[ghstack-poisoned]
