+++ torch/onnx/_internal/fx/dynamo_graph_extractor.py
@@ -194,8 +194,6 @@ class DynamoExport(exporter.FXGraphExtractor):
             *model_args,
             tracing_mode=fx_mode,
             fake_mode=fake_mode,  # type: ignore[arg-type]
-            aten_graph=fake_mode
-            is not None,  # TODO: Tracked by https://github.com/pytorch/pytorch/issues/105467
             **model_kwargs,
         )
         del graph_guard  # Unused
