+++ torch/onnx/_internal/fx/passes/type_promotion.py
@@ -32,7 +32,7 @@ from torch.fx.node import Node  # noqa: F401
 from torch.onnx._internal import _beartype
 from torch.onnx._internal.fx import _pass, diagnostics
 from torch.utils import _python_dispatch, _pytree
-
+from torch._dispatch.python import enable_python_dispatcher
 logger = logging.getLogger(__name__)
 
 # TODO(bowbao): move to type utils.
@@ -1613,19 +1613,6 @@ class InsertTypePromotion(_pass.Transform):
             diagnostic_context, module, type_promotion_table or TypePromotionTable()
         )
 
-    def _detect_fake_mode(self) -> Optional[fake_tensor.FakeTensorMode]:
-        """Detect fake mode from the graph.
-
-        Scan through all nodes in graph and their meta['val'] to detect fake mode.
-        """
-        fake_tensors = []
-        for node in self.module.graph.nodes:
-            try:
-                fake_tensors.append(_fake_tensor_from_node_val(node))
-            except RuntimeError:
-                continue
-        return torch._dynamo.utils.detect_fake_mode(fake_tensors)
-
     def _fetch_fake_args(self) -> Sequence[Optional[fake_tensor.FakeTensor]]:
         """Fetch fake args from fx graph.
 
@@ -1661,12 +1648,10 @@ class InsertTypePromotion(_pass.Transform):
         )
         assert not kwargs, "`kwargs` is not supported"
 
-        fake_tensor_mode = self._detect_fake_mode() or fake_tensor.FakeTensorMode(
-            allow_non_fake_inputs=True
-        )
         fake_args = self._fetch_fake_args()
+        fake_mode = self.fake_mode or torch._dynamo.utils.detect_fake_mode(fake_args)
 
-        with fake_tensor_mode, fx_traceback.preserve_node_meta():
+        with fake_mode, fx_traceback.preserve_node_meta():
             self.interpreter.run(*fake_args)
 
         return self.module
